Add `combine` options for visa and dinersclub
The `possibleProviders` function now accepts an options hash, including
the `combineVisa` and `combineDinersClub` options.  If enabled (or if
not specified), it will combine "visa" and "visa electron", as well as
all the "dinersClub*" varients into just "visa" and "dinersClub",
respectively.  I believe this is what most people will expect this
library to do.